rename and move openapi_client to api/openapi_generated
