+++ .github/workflows/push.yml
@@ -31,7 +31,7 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: generated-client-${{ matrix.python-version }}
-        path: python_sdk/src/openapi_client.py
+        path: python_sdk/src/api/openapi_generated.py
 
   build:
     needs: [generateStub]
