+++ python_sdk/.flake8
@@ -8,7 +8,7 @@ exclude =
     .pytest_cache,
     build,
     dist,
-    openapi_client.py,
+    openapi_generated.py,
     venv,
     __init__.py,
     setup.py
