+++ python_sdk/.gitignore
@@ -173,4 +173,4 @@ cython_debug/
 # PyPI configuration file
 .pypirc
 
-openapi_client.py
+openapi_generated.py
