+++ python_sdk/mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
 exclude = build
 
-[mypy-src.openapi_client.*]
+[mypy-src.api.openapi_generated]
 ignore_errors = True
