+++ python_sdk/scripts/generate_stub.bat
@@ -11,7 +11,7 @@ REM Install required packages
 pip install datamodel-code-generator
 
 REM Generate client code
-datamodel-codegen --input ../VizQLDataServiceOpenAPISchema.json --output-model-type pydantic_v2.BaseModel --input-file-type openapi --output src/openapi_client-raw.py --use-annotated --allow-population-by-field-name
+datamodel-codegen --input ../VizQLDataServiceOpenAPISchema.json --output-model-type pydantic_v2.BaseModel --input-file-type openapi --output src/api/openapi_generated-raw.py --use-annotated --allow-population-by-field-name
 
 echo Starting post-processing...
 python scripts/post_process.py
